APIs/billingo.hu/3.0.7/python-aiohttp-client/tests/test_product_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.client_error_response import ClientErrorResponse
from openapi_server.models.product import Product
from openapi_server.models.product_list import ProductList
from openapi_server.models.server_error_response import ServerErrorResponse
from openapi_server.models.validation_error_response import ValidationErrorResponse


pytestmark = pytest.mark.asyncio

async def test_create_product(client):
    """Test case for create_product

    Create a product
    """
    body = {"unit":"unit","general_ledger_number":"general_ledger_number","name":"name","vat":"0%","comment":"comment","currency":"AUD","general_ledger_taxcode":"general_ledger_taxcode","id":6,"net_unit_price":1.4658129}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='POST',
        path='/v3/products',
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_product(client):
    """Test case for delete_product

    Delete a product
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='DELETE',
        path='/v3/products/{id}'.format(id=56),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_product(client):
    """Test case for get_product

    Retrieve a product
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v3/products/{id}'.format(id=56),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_product(client):
    """Test case for list_product

    List all product
    """
    params = [('page', 56),
                    ('per_page', 25)]
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v3/products',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_update_product(client):
    """Test case for update_product

    Update a product
    """
    body = {"unit":"unit","general_ledger_number":"general_ledger_number","name":"name","vat":"0%","comment":"comment","currency":"AUD","general_ledger_taxcode":"general_ledger_taxcode","id":6,"net_unit_price":1.4658129}
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='PUT',
        path='/v3/products/{id}'.format(id=56),
        headers=headers,
        json=body,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

